Add files via upload
Went through the backprop for the probs and dprobs, had to use good 'olm chain rule for the derivative of the probs, reason is that dlogprobs in excercise 1a is the outside function as logprobs = probs.log() and we need to find the derivative of probs using the derivative of logprobs.  

Note: This is not for making our model better, but for an increase understanding of what back propagation is doing behind the scenes.